+++ examples/006_wakefields/000b_sum_wakefields.py
@@ -27,7 +27,7 @@ wf_df = xf.read_headtail_file(
     wake_file=wake_table_filename,
     wake_file_columns=wake_file_columns
 )
-wf0 = xf.Wakefield.from_table(
+wf0 = xf.WakefieldFromTable(
     wf_df,
     use_components=['dipole_x', 'dipole_y'],
     zeta_range=(-0.5*bucket_length_m, 0.5*bucket_length_m),
