+++ xfields/beam_elements/wakefields/__init__.py
@@ -1,4 +1,4 @@
 from .wakefield import WakeComponent, ResonatorWake
 from .wakefield import Wakefield
 from .read_headtail_table import read_headtail_file
-from .wake_field_from_table import WakefieldFromTable
+from .wakefield_from_table import WakefieldFromTable
